Adding mpi_base module
