+++ .circleci/config.yml
@@ -175,12 +175,12 @@ jobs:
                 /hpx/source \
                 -G "Ninja" \
                 -DCMAKE_BUILD_TYPE=Debug \
+                -DHPX_MPI_WITH_FUTURES=Off \
                 -DHPX_WITH_MALLOC=system \
                 -DHPX_WITH_GIT_COMMIT=${CIRCLE_SHA1} \
                 -DHPX_WITH_GIT_BRANCH="${CIRCLE_BRANCH}" \
                 -DHPX_WITH_GIT_TAG="${CIRCLE_TAG}" \
                 -DHPX_WITH_TOOLS=On \
-                -DHPX_MPI_WITH_FUTURES=Off \
                 -DCMAKE_CXX_FLAGS="-fcolor-diagnostics" \
                 -DHPX_WITH_TESTS_HEADERS=On \
                 -DHPX_WITH_COMPILER_WARNINGS=On \
@@ -424,7 +424,9 @@ jobs:
           name: Installing HPX for build unit tests
           command: |
               # Disable documentation just for this step
-              cmake -DHPX_MPI_WITH_FUTURES=Off -DHPX_WITH_DOCUMENTATION=OFF .
+              cmake \
+              -DHPX_MPI_WITH_FUTURES=Off \
+              -DHPX_WITH_DOCUMENTATION=Off .
               ninja -j1 install
       - run:
           name: Building Unit Tests
