+++ .github/workflows/macos.yml
@@ -26,8 +26,8 @@ jobs:
               -H. \
               -Bbuild \
               -GNinja \
+              -DHPX_MPI_WITH_FUTURES=Off \
               -DCMAKE_BUILD_TYPE=Debug \
-              -DHPX_MPI_WITH_FUTURES=OFF \
               -DHPX_WITH_EXAMPLES=ON \
               -DHPX_WITH_TESTS=ON
     - name: Build
