+++ .github/workflows/windows.yml
@@ -23,7 +23,7 @@ jobs:
     - name: Configure
       shell: bash
       run: |
-          cmake . -Bbuild -G'Visual Studio 16 2019' -DCMAKE_BUILD_TYPE=Debug -DCMAKE_TOOLCHAIN_FILE='C:/projects/vcpkg/scripts/buildsystems/vcpkg.cmake' -DHPX_WITH_PSEUDO_DEPENDENCIES=OFF -DHPX_WITH_EXAMPLES=ON -DHPX_WITH_TESTS=ON -DHPX_WITH_DEPRECATION_WARNINGS=OFF -DHPX_WITH_TESTS_MAX_THREADS_PER_LOCALITY=2 -DHPX_MPI_WITH_FUTURES=OFF
+          cmake . -Bbuild -G'Visual Studio 16 2019' -DHPX_MPI_WITH_FUTURES=Off -DCMAKE_BUILD_TYPE=Debug -DCMAKE_TOOLCHAIN_FILE='C:/projects/vcpkg/scripts/buildsystems/vcpkg.cmake' -DHPX_WITH_PSEUDO_DEPENDENCIES=OFF -DHPX_WITH_EXAMPLES=ON -DHPX_WITH_TESTS=ON -DHPX_WITH_DEPRECATION_WARNINGS=OFF -DHPX_WITH_TESTS_MAX_THREADS_PER_LOCALITY=2
     - name: Build
       shell: bash
       run: |
