+++ hpx/plugins/parcelport/mpi/header.hpp
@@ -13,8 +13,7 @@
 #if defined(HPX_HAVE_NETWORKING) && defined(HPX_HAVE_PARCELPORT_MPI)
 
 #include <hpx/assertion.hpp>
-#include <hpx/plugins/parcelport/mpi/mpi.hpp>
-#include <hpx/plugins/parcelport/mpi/mpi_environment.hpp>
+#include <hpx/mpi_base.hpp>
 #include <hpx/runtime/parcelset/parcel_buffer.hpp>
 
 #include <array>
