+++ hpx/plugins/parcelport/mpi/sender.hpp
@@ -15,7 +15,7 @@
 #include <hpx/assertion.hpp>
 #include <hpx/synchronization/spinlock.hpp>
 
-#include <hpx/plugins/parcelport/mpi/mpi_environment.hpp>
+#include <hpx/mpi_base.hpp>
 #include <hpx/plugins/parcelport/mpi/sender_connection.hpp>
 #include <hpx/plugins/parcelport/mpi/tag_provider.hpp>
