+++ libs/command_line_handling/include/hpx/command_line_handling/command_line_handling.hpp
@@ -8,11 +8,10 @@
 #define HPX_UTIL_COMMAND_LINE_HANDLING_HPP
 
 #include <hpx/config.hpp>
-#include <hpx/functional/function.hpp>
+#include <hpx/functional.hpp>
 #include <hpx/program_options.hpp>
-#include <hpx/runtime_configuration/runtime_configuration.hpp>
-#include <hpx/runtime_configuration/runtime_mode.hpp>
-#include <hpx/util/manage_config.hpp>
+#include <hpx/runtime_configuration.hpp>
+#include <hpx/util.hpp>
 
 #include <cstddef>
 #include <string>
@@ -21,9 +20,6 @@
 
 ///////////////////////////////////////////////////////////////////////////////
 namespace hpx { namespace util {
-    namespace detail {
-        HPX_EXPORT bool check_mpi_environment(runtime_configuration const& cfg);
-    }
 
     ///////////////////////////////////////////////////////////////////////////
     // Helper functions for retrieving command line options (with error
