+++ libs/logging/include/hpx/logging.hpp
@@ -162,6 +162,7 @@ struct bootstrap_logging
 template <typename T>
 bootstrap_logging const& operator<<(bootstrap_logging const& l, T const& t)
 {
+    // NOLINTNEXTLINE(bugprone-branch-clone)
     LBT_(info) << t;
     LPROGRESS_ << t;
     return l;
