+++ libs/mpi_base/docs/index.rst
@@ -1,0 +1,18 @@
+..
+    Copyright (c) 2020 The STE||AR-Group
+
+    SPDX-License-Identifier: BSL-1.0
+    Distributed under the Boost Software License, Version 1.0. (See accompanying
+    file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+.. _libs_mpi_base:
+
+========
+mpi_base
+========
+
+TODO: High-level description of the library.
+
+See the :ref:`API reference <libs_mpi_base_api>` of this module for more
+details.
+
