+++ libs/mpi_base/include/hpx/mpi_base/mpi.hpp
@@ -11,19 +11,19 @@
     defined(HPX_HAVE_LIB_MPI)
 
 #if defined(__clang__)
-#  pragma clang diagnostic push
-#  pragma clang diagnostic ignored "-Wcast-qual"
-#elif defined (__GNUC__)
-#  pragma GCC diagnostic push
-#  pragma GCC diagnostic ignored "-Wcast-qual"
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored "-Wcast-qual"
+#elif defined(__GNUC__)
+#pragma GCC diagnostic push
+#pragma GCC diagnostic ignored "-Wcast-qual"
 #endif
 
 #include <mpi.h>
 
 #if defined(__clang__)
-#  pragma clang diagnostic pop
-#elif defined (__GNUC__)
-#  pragma GCC diagnostic pop
+#pragma clang diagnostic pop
+#elif defined(__GNUC__)
+#pragma GCC diagnostic pop
 #endif
 
 #endif
