+++ libs/util/CMakeLists.txt
@@ -10,6 +10,7 @@ set(util_headers
   hpx/util/detail/reserve.hpp
   hpx/util/await_traits.hpp
   hpx/util/calculate_fanout.hpp
+  hpx/util/get_entry_as.hpp
   hpx/util/get_and_reset_value.hpp
   hpx/util/get_entry_as.hpp
   hpx/util/insert_checked.hpp
