+++ plugins/parcelport/mpi/CMakeLists.txt
@@ -16,12 +16,10 @@ if(HPX_WITH_NETWORKING AND HPX_WITH_PARCELPORT_MPI)
   add_parcelport(mpi
     STATIC
     SOURCES
-      "${PROJECT_SOURCE_DIR}/plugins/parcelport/mpi/mpi_environment.cpp"
       "${PROJECT_SOURCE_DIR}/plugins/parcelport/mpi/parcelport_mpi.cpp"
     HEADERS
       "${PROJECT_SOURCE_DIR}/hpx/plugins/parcelport/mpi/header.hpp"
       "${PROJECT_SOURCE_DIR}/hpx/plugins/parcelport/mpi/locality.hpp"
-      "${PROJECT_SOURCE_DIR}/hpx/plugins/parcelport/mpi/mpi_environment.hpp"
       "${PROJECT_SOURCE_DIR}/hpx/plugins/parcelport/mpi/receiver.hpp"
       "${PROJECT_SOURCE_DIR}/hpx/plugins/parcelport/mpi/receiver_connection.hpp"
       "${PROJECT_SOURCE_DIR}/hpx/plugins/parcelport/mpi/sender.hpp"
@@ -39,6 +37,7 @@ if(HPX_WITH_NETWORKING AND HPX_WITH_PARCELPORT_MPI)
       hpx_filesystem
       hpx_hardware
       hpx_memory
+      hpx_mpi_base
       hpx_performance_counters
       hpx_plugin
       hpx_program_options
@@ -48,27 +47,6 @@ if(HPX_WITH_NETWORKING AND HPX_WITH_PARCELPORT_MPI)
       hpx_threadmanager
       hpx_timing
       hpx_topology
-      hpx_util
-      hpx_dependencies_boost
-      Mpi::mpi
-    INCLUDE_DIRS
-      ${PROJECT_SOURCE_DIR}
-    FOLDER
-      "Core/Plugins/Parcelport/MPI")
-elseif(HPX_MPI_WITH_FUTURES)
-  hpx_debug("add_parcelport_mpi_module (reduced)")
-  include(HPX_AddParcelport)
-  add_parcelport(mpi
-    STATIC
-    SOURCES
-      "${PROJECT_SOURCE_DIR}/plugins/parcelport/mpi/mpi_environment.cpp"
-    HEADERS
-      "${PROJECT_SOURCE_DIR}/hpx/plugins/parcelport/mpi/mpi_environment.hpp"
-    DEPENDENCIES
-      hpx_config
-      hpx_command_line_handling
-      hpx_runtime_configuration
-      hpx_synchronization
       hpx_util
       hpx_dependencies_boost
       Mpi::mpi
